backend/pyproject.toml
[project]
name = "market_sentiment_backend"
version = "0.1.0"
description = "Flask backend for visual sentiment analysis and market prediction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0",
    "flask-cors>=4.0",
    "python-dotenv>=1.0",
    "torch>=2.4",
    "torchvision>=0.19",
    "transformers>=4.40",
    "pillow>=10.0",
    "opencv-python>=4.10",
    "easyocr>=1.7",
    "pytesseract>=0.3.10",
    "yfinance>=0.2.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "sqlalchemy>=2.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "black>=24.8",
    "ruff>=0.6"
]


